s2_calibrate now handles multi-segment files
Continuing to address Issue #60.

s2_calibrate now takes a data folder instead of a data file and analyzes
all files in that folder with the suffix rateints (note this will not
work if an instrument or observing mode does not produce rateints after
the Stage 1 pipeline is run).

Further testing is needed to determine whether the skip flags are later
overridden. Successfully tested on MIRI LRS data, and testing for
NIRSpec and NIRCam to come soon. The main question is whether the
correct steps will be run for each instrument.