s2_calibrate tested on MIRI LRS and NIRSpec FS
Continuing to address Issue #60.

S2 now works well for MIRI and NIRSpec
which each have template ecf files now, but NIRCam fails with an error
message produced during extract_1d which says the following:
'1' is not of type 'integer'
which seems to be in reference to the dither pattern or position. More
troubleshooting to come.