Skip booleans do not work
Continuing to address Issue #60.

Code tested with non-default skip boolean flags, and they are indeed
overridden as I had worried. I will work on fixing these options next.
While this can be done with a .cfg file, these are now deprecated and
ASDF files are preferred. Ideally we wouldn't need two separate control
files though, so I will continue looking for ways to do things within
python.

Note that if someone wants to change one of the default functions, this
can be done by setting self.fname equal to some different function that
takes the same inputs as the original "fname" function. If more inputs are
needed, we will need to make a copy of calwebb_spec2.py in order to add
the extra arguments when calling the function.