Skip booleans are now respected.
Continuing to address Issue #60.

I needed to change how the pipeline was called from a self.run() call to
a self() call. The former takes a cfg file and overwrites the set
booleans, but the latter does not take a cfg file and respects the set
booleans.

Also changed the comments beside some flags in the ecf templates as not
all steps are run for TSO observations.

Still to come is debugging the dither_points issue for NIRCam and
testing integration with Stages 3+. In addition, there are several other
tasks listed in Issue #35 which may or may not be specific to each
instrument. Changing dispersion axis trimming is likely still not
working if it wasn't working as of commit b623eb2