Further cleaning S2 code
Now the suffix for S2 is input in the ecf file rather than requiring it
be rateints. Also moved the code that allows files to be written outside
of the Eureka project folder to the util.py file to avoid duplicated
work/code. The flat-field subtraction step is now set to be skipped in
the NIRSpec ecf template according to comments from Aarynn. Also made it
possible for any instrument to have slit and waverange parameters set
in the ecf.

Will need to change S3+ code to use the new util.py code. Will improve
the integration between S2 and S3 (and later S3 and S4) as my next step.