Improving the integration between S2 and S3.
This may be the last step in addressing Issue #60.

Now pass the S2 metadata object into the S3 script when calling S2 and
S3 sequentially. I take advantage of this to figure out where exactly
the S2 outputs are. Otherwise, if running S3 without running S2, you
will need to put the exact folder that contains the outputs from S2 (be
they from Eureka or the JWST pipeline). Also implemented a feature where
the datadir in the copied ecf file is overwritten with whatever was
actually used to have a more useful ecf file in the case that you want
to look back and see which particular S2 outputs were used in that S3
run without needing to look at the metadata file.

Also fixed a bug mentioned in Issue #36 where multiple MIRI FITS
segments in S3 would cause the xwindow and ywindow to keep swapping back
and forth.

S2 -> S3 integration has been well tested for MIRI and NIRCam, but I
can't test NIRSpec myself as it overflows my RAM. There is no reason to
assume that NIRSpec would fail though.