S3 can now search for a metadata file
S3 can now search for a metadata file. This is helpful if not running S2
and S3 sequentially - Eureka will now look for the metadata output from
an S2 step, and if it isn't able to find it, it will issue a warning
that it could not find a metadata object and is proceeding assuming the
data was produced by the JWST pipeline.