Updating S4 to take the outputs from updated S3
This includes adopting the namechanges of inputdir and outputdir as
well as a bunch of additional code to permit S4 to handle the changes to
S3 which permitted multiple apertures to be considered. The user has the
option of only considering one aperture (either the only one in the
folder or the newest one according the file filename) or running S4 on
all apertures from one run of S3.

Running S2-S4 now well tested for MIRI and NIRCam (but I'm still having
RAM issues with NIRSpec). Drift correction doesn't work for the MIRI
simulated data, but these spectra are very smooth (blackbody), which
might be the source of the problem. I will try making some MIRI
simulations with a stellar model in the next few days to see if that
fixes things.

More documentation is needed for all stages on the readthedocs page.
This will be among the next things I tackle (or at least make a first
attempt at).