Updating the run_eureka.py in the S3 demo folder
This file previously had the wrong format for the s4 call. This file is
now updated. I debated deleting this file, but it is worth keeping as
not everyone will want to start with S2 so they may look at this file
first or want to know how to proceed without running S2 first.