Tackled much of the docstring writing for S2-S4
I've gone through most of the Eureka! files and added/clarified the
functions' docstrings to allow for nice API on the readthedocs page. The
main thing that is missing is the eureka/lib folder which has quite a
lot of files in it. These are less likely to be tweaked or directly
examined though, so I'll save those for another time.