New option: subtract continuum before x-correlate
This commit at least partially addresses Issue #66.

Added the option to S4 to subtract the continuum from the spectra before
cross-correlating with a reference spectrum to search for spectral
drift/jitter. This can help since continuum variations can mess with the
cross-correlation step.

Also setting correctDrift to False as default in the S4_template.ecf
and added a note that this step is not recommended for simulated data.