Generalized S2 to allow imaging mode observations.
Resolves Issue #70

Imaging mode observations needed the Image2Pipeline instead of the
Spec2Pipeline, but much of the previously written code still works well.
Moved some of the code around to avoid repeated code and to make S2
easier to use and more like the other stages. Changed skip_resample_spec
to skip_resample in the spectral templates to keep the names consistent
between imaging and spectral templates. Added a new imaging template to
show what steps are relevant and which are skipped for TSO mode
observations. Code tested on MIRI LRS spectroscopy and SUB64 imaging
simulations.