Adding a new wfc3 instrument file
I've implemented as much of wfc3_1reduce.py into a new file called
wfc3.py which tries to keep as much of the same structure as the JWST
instruments while also trying to keep as much of the WFC3 code as
possible. Where possible, WFC3 uses the same code as JWST (e.g.
background subtraction, optimal extraction) to reduce duplicated code.

The main exception that I ran into is that Eureka works one file at a
time while the WFC3 code used to open all files at once. This allowed
full time-axis sigma clipping but also requires more RAM which is
likely unrealistic for JWST. So I have decided to keep WFC3 analyses
similar to JWST ones which means there is no longer full time-axis sigma
clipping for individual pixels. The other impact is that currently the
reference files for each scan direction are the first files (because
they are read first). This could be improved in the future by
temporarily moving the desired reference frames to the start of the
segment_list array and then re-sorting everything afterwards.

The next step will be to call wfc3.py from s3_reduce.py and add some
empty functions to the other instruments (e.g. correct_drift2D which
might be worth implementing for those other instruments in the future).