Significant bug fixes for wfc3.py
There's a lot that went into this, but the wfc3.py file now gives a
fairly clean transit light curve for some real WFC3 G102 observations.
One of the major edits was properly reading in the FITS files to get
all of the reads, rather than just the last.

Temporarily removing the need for the leapsecond corrections to get
things working for a first run-through test.