Making S5 iterate over all apertures too
S4 has the option to run iteratively on all the apertures considered in
S3, and now S5 can do the same. This will allow for nice end-to-end
tests of aperture sizes with minimal manual labour.

Also made some changes so that the continuity of information is
maintained between S4 and S5 (e.g. S5 log is appended to old S4 log,
etc.).