Calling S2-5 under main func for multiprocessing
I was having issues where calling pool.map_async would cause each
process to start from the very beginning of run_eureka.py. For some
reason this is resolved by calling S2-5 under an
"if __name__=='__main__':"
statement. It is unclear why I would only encounter this on Mac, and
potentially why others on MacOS have also not encountered the issue.
For now though, this is a benign change which should work for all