Separate eclipse model, new phase curve model
Separated the transit and eclipse models to make things cleaner when
modelling eclipses (don't need to specify a uniform limb darkening for
eclipse modelling anymore). To fit a transit, one can use eclipse in
run_myfuncs

Also added a new phase variations model that uses the same formulation
as SPCA. To use phase variations, one uses phase in run_myfuncs. The
relevant phase variation parameters are A, B, C, and D which are
coefficients for the first order cos and sin and second order cos and
sin.

Made it optional to fit for time of eclipse directly, and can instead
fit for time of transit, period, ecc, and/or w.

Made testing a model separate from meta.testing_S5 since I wanted to be
able to test my new phase variation models with only one wavelength
without injecting the exponential ramp as well.