Parallel processing in emcee and dynesty fitters
To make this happen, I needed to make a smaller number of tweaks.
First, the lc object no longer stores the log as this cannot be pickled.
Second, I had to remove all reload statements in Stage 5 as I could not
pickle reloaded classes (pickling is required for multiprocessing).
Finally, BatmanModels no longer store self.ld_func - if this is needed,
I can write more code later to allow ld_func to be pickled, but it is
currently unused and didn't seem worth the effort.

The number of threads is set in the S5 ecf, and if unset will default to
not running in parallel.

Also fixing merge conflict from PR #153. Also suppressing
EXOCTK_DATA warning (Issue #177).