niriss.py
Moved a bunch of functions outside of previous functions to be able to use elsewhere. Included a routine for a simple box extraction (used in the optimal extraction). Included `flag_bg` which is just a wrapper for `fit_bg`, since I complete the cosmic ray removal simultaneously with the background fitting