verbose controls print not log, adds S3,4,6 tables
This commit adds astropy table outputs from S3, S4, and S6 to help
prepare submissions for the ERS data challenge and in gerenal help
transitions to non-Eureka code. My next commit will add similar outputs
to S5.

This commit changes the verbose option in the ecf to control what gets
printed to the screen while making sure the full outputs are still
printed to the log.