Adds S5 table output, logs lnprob during lsq
This commit adds an astropy table output from S5 similar to that from S4
but which includes the model and residuals from the model as new columns
and uses the normalized flux and uncertainties (and instead uses the
fitted uncertainties if present).

This commit also logs the lnprob during an lsq fit and logs the lnprob
before and after every fitter.