starry allows all unshared; crashes for many ch
The code works with 2 channels, but for 37 I get this error message:
"fatal error: bracket nesting level exceeded maximum of 256".

Reading online, it seems we'll need to convert c0, c0_1, etc. into a
single vector named c instead which should help (and the same for all
parameters). I was initially going to take that route, but wanted to
keep things more similar to the non-gradient-based models.