Allowing a meta.time_file for user-input times
As of right now, to fix bugs with the time arrays we need to directly
edit the code. This new patch will allow users to provide a file with
the time array that they would like. I've made it so that separate rows
or whitespace columns are allowed for the time file.