Fixing memory overflow with S3 NIRISS
Also commenting out interpolating_image calls which are extremely slow.
We should discuss whether we want to keep this function and whether
there's anything we can do to speed it up.

The util.check_nans feature doesn't seem to be working as expected
either as there are 1449536 NaNs in the simualted data for the ERR and
V0 arrays.

And after that, I eventually get:
```
Traceback (most recent call last):
  File "/Users/tjbell1/Research/JWST/ERS/Eureka/ecfs_NIRISS/run_eureka.ecf", line 23, in <module>
    s3_meta = s3.reduce(eventlabel)
  File "/Users/tjbell1/Research/JWST/ERS/Eureka/ecfs_NIRISS/../src/eureka/S3_data_reduction/s3_reduce.py", line 78, in reduce
    s3_reduce_niriss.reduce(eventlabel,
  File "/Users/tjbell1/Research/JWST/ERS/Eureka/ecfs_NIRISS/../src/eureka/S3_data_reduction/s3_reduce_niriss.py", line 249, in reduce
    data = inst.fit_bg(data, meta, readnoise=meta.readnoise,
  File "/Users/tjbell1/Research/JWST/ERS/Eureka/ecfs_NIRISS/../src/eureka/S3_data_reduction/niriss.py", line 150, in fit_bg
    box_mask = dirty_mask(data.medflux.values,
  File "/Users/tjbell1/Research/JWST/ERS/Eureka/ecfs_NIRISS/../src/eureka/S3_data_reduction/niriss_extraction.py", line 127, in dirty_mask
    order3[:, i] = img[s:e, i]
ValueError: could not broadcast input array from shape (0,) into shape (60,)
```