Speeding up optspec and meta.record_ypos
Using parallelized code to speed up meta.record_ypos which had been
slow. Also moved code from s3_reduce to source_pos and optspex to
better modularize the code.

Also added the ability to set a limit on how many figures are made of a
certain type (currently only applies to Figs 3301, 3302, 3304, and 3501
which are the most annoying figures to get tonnes of.