Properly handling r ~< 0 for fig6301
Before r < 0 was allowed to be shown which would break the secondary
H(r) y-axis. This patch addresses that by forcing r>=0 for fig 6301. I
also ran into issues with overlapping tick marks near r=0, so I added
some extra code to avoid that. It isn't the nicest, but it's what I
could think to do.