Protecting logedit.py from brief drive disconnects
If the drive to which the log is being written is briefly disconnected,
writing to the log later then causes the whole run to crash. This has
happened several times for me, and this tweak should fix that by trying
to reopen the log and just not logging if it is unable to write to the
log.