Modularized differentiable fns, working on merging
Took a lot of effort, but now the differentiable functions are
modularized. Excitingly, the differentiable functions appear to work
with lsq, giving the exact same results as the non-differentiable
functions with lsq.

As of right now though, using exoplanet instead of lsq gives quite
different results, so it's possible there's some difference between the
eval and non-eval parts of the code