Solve for Mp so that starry won't overwrite period
Starry only lets you set the orbital period OR the semi-major axis,
and requires that you provide the planetary mass so that it can compute
the other parameter. I now instead set the planetary mass such that
starry's computed semi-major axis and/or period agree with the value in
the fit.