Added new centroiding method mgmc and fixed bugs.
Stage3 is having an issue with the method used for centroiding.
The method currently used is fgc it is also the cuase of the issue.
When running SW photometry fgc struggles with centroiding.
So I added a new method mgmc.
This new method is called mgmc and uses gaussian centroiding.
This allows for the detection of a centroid location x, y.
Along with 1-sig gaussian widths sx, sy.
This enables the code to correclty find the source in the calints files.
Allowing for a more precise and fucntioning method for SW photometry.
Bugs were also addressed such as masking NANs and plotting figures.