Changing the MAD computation direction
Resolves #304

Previously, the MAD was computed along the wavelength axis and then
an average value was computed. Minimizing this 2D MAD is essentially
minimizing noise in the spectral direction. Instead, we typically want
to minimize noise in the temporal direction, so I've rotated the MAD
calculation so that it is computed along the time axis for each
wavelength, and then averaged. I also separately handle each WFC3 scan
direction to avoid issues there.