Removing 390Hz noise from LRS SLITLESSPRISM data
Some MIRI subarrays exhibit 390Hz noise in the pixels within a group.
We roughly knew about this from the LRS TSO commissioning
paper, but Michael Ressler did some great investigative work and showed
that there is a clear 390Hz noise in the group-level data which has a
fixed waveform but has a varying phase. I did some testing myself and
determined the waveform for the LRS SLITLESSPRISM subarray and wrote
code to find the phase. Interestingly, I found that the phase remains
constant within an integration, and while the phase varies a bit between
integrations, I noticed that it has cyclical variations with a period of
8 integrations (in at least one dataset). I couldn't figure out a clean
way to deterministically compute the phase however, so I just fit it in
my added code.

I had hoped that this would improve the high-frequency noise I've
observed in LRS spectra (the noise is in the final transit/eclipse
spectra). However, I have found that the 390Hz step seems to have nearly
zero impact on the transit/eclipse spectrum, while the related
MIRI-specific GLBS code I added at the same time also has little impact.
That said, I don't think I'm the only one who would want to investigate
this effect, so I am still contributing these edits to allow others to
investigate the impact on their data and allow for potential further
improvements.

I will note that this code was only built with the LRS SLITLESSPRISM
subarray in mind and has been tested on several such observations, but a
few other subarrays also exhibit a similar noise pattern. I don't have
any such observations to test the code on, so for now this step should
only be applied to LRS SLITLESSPRISM data.