Fixing GPModel for fitting single lightcurve
The GPModel.py file was long neglected and had numerous bugs in it, some
of which were from updates to other models. I've significantly cleaned up
the code and changed the fitted GP variables. So far I've tested this
with success on a photometric lightcurve.

There are still known bugs when doing multwhite fits, and there may well
be bugs in shared fits. As a result, I've currently prohibited multwhite
and shared fits with the GP. I'll work on getting those enabled next.