Messy, temporary code to enable the fitting multiple planets
This temporary, hack-y code is specifically for the TRAPPIST-1b+c phase curve which also has a transit of TRAPPIST-1g. As a result, the code is edited in very specific ways that will only work for those observations. I will need to come back and generalize this code to work for multi-planet systems in general.