Changing starry model to no longer require Ms
The requirement for Ms was always a nuisance since it could potentially crash the code if Ms was larger than allowed. The only reason it was added was so that starry would be able to accept both a/Rs and Porb, but Everett Schlawin pointed out that we can just set Mp=0 and then solve for Ms which is way easier and removes that nuisance parameter.