Adding _ch to channel tag and _pl to planet tag in S5 variable names
Also added support for ecosw and esinw in batman+POET models (not yet for starry models).

Still need to fully integrate _pl tags, especially for starry. Still need to do a bunch of testing.

Note, this change is going to break backwards compatibility for S6 running on old S5 outputs, but that isn't a huge deal since S6 isn't critical and we're approaching v1.0