Added code to auto-generate pixel priors from one EPF line
This way users can specify a prior for all pixels without having to add a tonne of priors to their EPF. But users can still provide separate priors for each pixel if they desire; any pixel not specified in the EPF will fall back to the 0th pixel. Also made sure that this works in the case of fits where the pixels are shared between fits.