Adding a second version of Fig 5501 corner plot for pixel mapping
The new version shows Ylm values instead of pixel values, and this new plot is in addition to the default version (not instead of it).

This code works great for single-lightcurve fits, but currently it is not going to work correctly for multwhite or shared fits. I'll fix that next