Removing some scary ~bool instances
With the built-in boolean type, using the bitwise inversion (~ operator) gives scary and unexpected results. Specifically, doing ~False gives -1 (which is still Falsey) and ~True gives -2 (which is still Truthy). This isn't the case for numpy.bool_ type (and as a result ~np.array([False]) gives np.array([True])), but this is the case for the built-in bool type.