Add wavelengths parameter to PlanetParams and update model initializations
This is still a work in progress, with just the pieces I currently need implemented

The new meta.wavelengths parameter will allow for some fitted parameters to be shared between only "channels" with the same wavelength. An example usecase of this is fitting all MIRI imaging observations of TRAPPIST-1 planets, which include F1280W and F1500W observations. For those data, one may want the eclipse depth to be shared between all observations in the same filter, while the orbital parameters would be shared between all observations